requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-ollama
PyMuPDF
fpdf
beautifulsoup4
requests
python-multipart